pyproject.toml
[project]
name = "archive-agent"
version = "0.1.0"
description = "Archive Agent: Smart Indexer with RAG Engine"
authors = [
    {name = "Dr.-Ing. Paul Wilhelm", email = "paul@wilhelm.dev"}
]
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "openai (>=1.69.0,<2.0.0)",
    "qdrant-client (>=1.13.3,<2.0.0)",
    "typer (>=0.15.2,<0.16.0)",
    "pillow (>=11.1.0,<12.0.0)",
    "charset-normalizer (>=3.4.1,<4.0.0)",
    "nltk (>=3.8.2,<4.0.0)",
    "rich (>=14.0.0,<15.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.poetry.scripts]
archive-agent = "archive_agent.__main__:app"


[tool.poetry.group.dev.dependencies]
pytest = "^8.0"
pytest-cov = "^5.0"


[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-ra -q --cov=archive_agent --cov-report=term-missing"


[tool.coverage.run]
omit = [
    "*/__init__.py"
]

[tool.coverage.report]
omit = [
    "*/__init__.py"
]
